# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * DataClean – Excel/CSV Data Cleaning API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ColumnProfileInferredType = typeof ColumnProfileInferredType[keyof typeof ColumnProfileInferredType];


export const ColumnProfileInferredType = {
  string: 'string',
  integer: 'integer',
  float: 'float',
  datetime: 'datetime',
  boolean: 'boolean',
  categorical: 'categorical',
} as const;

export type ColumnProfileStats = { [key: string]: unknown };

export interface ColumnProfile {
  name: string;
  inferred_type: ColumnProfileInferredType;
  null_count: number;
  unique_count: number;
  total_count?: number;
  sample_values: string[];
  stats: ColumnProfileStats;
}

export interface TableInfo {
  start_row: number;
  end_row: number;
  start_col: number;
  end_col: number;
  /** @nullable */
  header_row?: number | null;
  confidence: number;
}

export interface TableDetectionResult {
  sheet_name: string;
  tables: TableInfo[];
}

export type SuggestionCategory = typeof SuggestionCategory[keyof typeof SuggestionCategory];


export const SuggestionCategory = {
  type_fix: 'type_fix',
  format: 'format',
  duplicate: 'duplicate',
  missing_value: 'missing_value',
  validation: 'validation',
  structure: 'structure',
} as const;

export type SuggestionStatus = typeof SuggestionStatus[keyof typeof SuggestionStatus];


export const SuggestionStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  edited: 'edited',
} as const;

export interface Suggestion {
  id: string;
  /** @nullable */
  row_index?: number | null;
  column_name: string;
  /** @nullable */
  original_value?: string | null;
  /** @nullable */
  proposed_value?: string | null;
  reason: string;
  category: SuggestionCategory;
  status: SuggestionStatus;
}

export type RecipeStepModule = typeof RecipeStepModule[keyof typeof RecipeStepModule];


export const RecipeStepModule = {
  structure: 'structure',
  type: 'type',
  format: 'format',
  dedup: 'dedup',
  missing: 'missing',
  validation: 'validation',
  nl: 'nl',
} as const;

export type RecipeStepParams = { [key: string]: unknown };

export interface RecipeStep {
  step_id: string;
  module: RecipeStepModule;
  description: string;
  params: RecipeStepParams;
}

export type ValidationRuleParams = { [key: string]: unknown };

export interface ValidationRule {
  rule_id: string;
  column_name: string;
  rule_type: string;
  description: string;
  params?: ValidationRuleParams;
  enabled: boolean;
}

export interface ValidationRuleList {
  rules: ValidationRule[];
}

export interface SessionInfo {
  session_id: string;
  original_filename: string;
  sheets: string[];
  /** @nullable */
  selected_sheet?: string | null;
  columns: ColumnProfile[];
  suggestions: Suggestion[];
  recipe_steps: RecipeStep[];
  created_at: string;
}

export interface UploadResult {
  session_id: string;
  sheets: string[];
  original_filename: string;
}

export interface FileUploadInput {
  file: Blob;
}

export interface RecipeFileInput {
  file: Blob;
}

export interface StructureDetectInput {
  sheet_name: string;
  /** @nullable */
  header_row?: number | null;
  /** @nullable */
  start_row?: number | null;
}

export type SuggestionUpdateInputStatus = typeof SuggestionUpdateInputStatus[keyof typeof SuggestionUpdateInputStatus];


export const SuggestionUpdateInputStatus = {
  accepted: 'accepted',
  rejected: 'rejected',
  edited: 'edited',
} as const;

export interface SuggestionUpdateInput {
  status: SuggestionUpdateInputStatus;
  /** @nullable */
  edited_value?: string | null;
}

export type BulkSuggestionInputStatus = typeof BulkSuggestionInputStatus[keyof typeof BulkSuggestionInputStatus];


export const BulkSuggestionInputStatus = {
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface BulkSuggestionInput {
  /** @nullable */
  category?: string | null;
  /** @nullable */
  column_name?: string | null;
  status: BulkSuggestionInputStatus;
}

export interface NLCommandInput {
  instruction: string;
}

export type NLCommandPreviewIntent = { [key: string]: unknown };

export type NLCommandPreviewSampleBeforeItem = { [key: string]: unknown };

export type NLCommandPreviewSampleAfterItem = { [key: string]: unknown };

export interface NLCommandPreview {
  preview_id: string;
  description: string;
  intent: NLCommandPreviewIntent;
  sample_before: NLCommandPreviewSampleBeforeItem[];
  sample_after: NLCommandPreviewSampleAfterItem[];
  affected_count: number;
  /** @nullable */
  clarification_needed?: string | null;
}

export interface NLCommandConfirmInput {
  preview_id: string;
}

export type PagedPreviewRowsItem = { [key: string]: unknown };

export type PagedPreviewOriginalRowsItem = { [key: string]: unknown };

export interface PagedPreview {
  rows: PagedPreviewRowsItem[];
  total_rows: number;
  page: number;
  page_size: number;
  columns: string[];
  original_rows?: PagedPreviewOriginalRowsItem[];
}

export interface IssueItem {
  suggestion_id: string;
  /** @nullable */
  row_index?: number | null;
  column_name: string;
  category: string;
  /** @nullable */
  original_value?: string | null;
  /** @nullable */
  proposed_value?: string | null;
  description: string;
  resolution: string;
}

export type IssuesReportSummary = { [key: string]: unknown };

export interface IssuesReport {
  session_id: string;
  issues: IssueItem[];
  generated_at: string;
  summary?: IssuesReportSummary;
}

export interface RecipeExport {
  session_id: string;
  steps: RecipeStep[];
  markdown: string;
}

export interface AnalysisResult {
  columns: ColumnProfile[];
  suggestions: Suggestion[];
}

export interface SuggestionList {
  suggestions: Suggestion[];
  total: number;
}

export interface ApplyResult {
  suggestion: Suggestion;
  applied: boolean;
}

export interface BulkApplyResult {
  updated_count: number;
  suggestions: Suggestion[];
}

export interface NLConfirmResult {
  applied: boolean;
  recipe_step: RecipeStep;
}

export interface RecipeApplyResult {
  new_session_id: string;
  steps_applied: number;
  steps_skipped: number;
}

export type ValidationRuleUpdateParams = { [key: string]: unknown };

export interface ValidationRuleUpdate {
  enabled?: boolean;
  params?: ValidationRuleUpdateParams;
  description?: string;
}

export interface DeleteResult {
  deleted: boolean;
}

export type GetSuggestionsParams = {
status?: GetSuggestionsStatus;
category?: GetSuggestionsCategory;
column_name?: string;
};

export type GetSuggestionsStatus = typeof GetSuggestionsStatus[keyof typeof GetSuggestionsStatus];


export const GetSuggestionsStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  edited: 'edited',
} as const;

export type GetSuggestionsCategory = typeof GetSuggestionsCategory[keyof typeof GetSuggestionsCategory];


export const GetSuggestionsCategory = {
  type_fix: 'type_fix',
  format: 'format',
  duplicate: 'duplicate',
  missing_value: 'missing_value',
  validation: 'validation',
  structure: 'structure',
} as const;

export type GetPreviewParams = {
page?: number;
page_size?: number;
show_original?: boolean;
};

export type ExportSessionParams = {
format?: ExportSessionFormat;
};

export type ExportSessionFormat = typeof ExportSessionFormat[keyof typeof ExportSessionFormat];


export const ExportSessionFormat = {
  xlsx: 'xlsx',
  csv: 'csv',
} as const;