# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DataClean – Excel/CSV Data Cleaning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sessions/{session_id}:
    get:
      operationId: getSession
      tags: [sessions]
      summary: Get session metadata
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "404":
          description: Session not found
    delete:
      operationId: deleteSession
      tags: [sessions]
      summary: Delete and purge a session from memory
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Session not found
  /sessions/{session_id}/detect-structure:
    post:
      operationId: detectStructure
      tags: [sessions]
      summary: Run table/header structure detection on a chosen sheet
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StructureDetectInput"
      responses:
        "200":
          description: Table detection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableDetectionResult"
        "404":
          description: Session not found
  /sessions/{session_id}/analyze:
    post:
      operationId: analyzeSession
      tags: [sessions]
      summary: Run type inference, formatting scan, dedup scan, and generate suggestions
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Analysis complete with column profiles and suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "404":
          description: Session not found
  /sessions/{session_id}/suggestions:
    get:
      operationId: getSuggestions
      tags: [suggestions]
      summary: Return the suggestion queue with optional filters
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, rejected, edited]
        - name: category
          in: query
          schema:
            type: string
            enum: [type_fix, format, duplicate, missing_value, validation, structure]
        - name: column_name
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Suggestion list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuggestionList"
        "404":
          description: Session not found
  /sessions/{session_id}/suggestions/{suggestion_id}:
    post:
      operationId: updateSuggestion
      tags: [suggestions]
      summary: Accept, reject, or edit a single suggestion
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - name: suggestion_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuggestionUpdateInput"
      responses:
        "200":
          description: Updated suggestion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplyResult"
        "404":
          description: Not found
...
  /sessions/{session_id}/validation-rules:
    get:
      operationId: getValidationRules
      tags: [sessions]
      summary: Get auto-generated validation rules for the session
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Validation rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationRuleList"
        "404":
          description: Session not found
  /sessions/{session_id}/validation-rules/{rule_id}:
    patch:
      operationId: updateValidationRule
      tags: [sessions]
      summary: Update or remove a validation rule
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - name: rule_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidationRuleUpdate"
      responses:
        "200":
          description: Updated rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationRule"
        "404":
          description: Not found
components:
  parameters:
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    ColumnProfile:
      type: object
      required: [name, inferred_type, null_count, unique_count, sample_values, stats]
      properties:
        name:
          type: string
        inferred_type:
          type: string
          enum: [string, integer, float, datetime, boolean, categorical]
        null_count:
          type: integer
        unique_count:
          type: integer
        total_count:
          type: integer
        sample_values:
          type: array
          items:
            type: string
        stats:
          type: object
          additionalProperties: true
    TableInfo:
      type: object
      required: [start_row, end_row, start_col, end_col, confidence]
      properties:
        start_row:
          type: integer
        end_row:
          type: integer
        start_col:
          type: integer
        end_col:
          type: integer
        header_row:
          type: ["integer", "null"]
        confidence:
          type: number
    TableDetectionResult:
      type: object
      required: [sheet_name, tables]
      properties:
        sheet_name:
          type: string
        tables:
          type: array
          items:
            $ref: "#/components/schemas/TableInfo"
    Suggestion:
      type: object
      required: [id, column_name, reason, category, status]
      properties:
        id:
          type: string
        row_index:
          type: ["integer", "null"]
        column_name:
          type: string
        original_value:
          type: ["string", "null"]
        proposed_value:
          type: ["string", "null"]
        reason:
          type: string
        category:
          type: string
          enum: [type_fix, format, duplicate, missing_value, validation, structure]
        status:
          type: string
          enum: [pending, accepted, rejected, edited]
    RecipeStep:
      type: object
      required: [step_id, module, description, params]
      properties:
        step_id:
          type: string
        module:
          type: string
          enum: [structure, type, format, dedup, missing, validation, nl]
        description:
          type: string
        params:
          type: object
          additionalProperties: true
    ValidationRule:
      type: object
      required: [rule_id, column_name, rule_type, description, enabled]
      properties:
        rule_id:
          type: string
        column_name:
          type: string
        rule_type:
          type: string
        description:
          type: string
        params:
          type: object
          additionalProperties: true
        enabled:
          type: boolean
    ValidationRuleList:
      type: object
      required: [rules]
      properties:
        rules:
          type: array
          items:
            $ref: "#/components/schemas/ValidationRule"
    SessionInfo:
      type: object
      required: [session_id, original_filename, sheets, columns, suggestions, recipe_steps, created_at]
      properties:
        session_id:
          type: string
        original_filename:
          type: string
        sheets:
          type: array
          items:
            type: string
        selected_sheet:
          type: ["string", "null"]
        columns:
          type: array
          items:
            $ref: "#/components/schemas/ColumnProfile"
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/Suggestion"
        recipe_steps:
          type: array
          items:
            $ref: "#/components/schemas/RecipeStep"
        created_at:
          type: string
          format: date-time
...
    StructureDetectInput:
      type: object
      required: [sheet_name]
      properties:
        sheet_name:
          type: string
        header_row:
          type: ["integer", "null"]
        start_row:
          type: ["integer", "null"]
    SuggestionUpdateInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, rejected, edited]
        edited_value:
          type: ["string", "null"]
...
    AnalysisResult:
      type: object
      required: [columns, suggestions]
      properties:
        columns:
          type: array
          items:
            $ref: "#/components/schemas/ColumnProfile"
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/Suggestion"
    SuggestionList:
      type: object
      required: [suggestions, total]
      properties:
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/Suggestion"
        total:
          type: integer
    ApplyResult:
      type: object
      required: [suggestion, applied]
      properties:
        suggestion:
          $ref: "#/components/schemas/Suggestion"
        applied:
          type: boolean
...
    ValidationRuleUpdate:
      type: object
      properties:
        enabled:
          type: boolean
        params:
          type: object
          additionalProperties: true
        description:
          type: string
    DeleteResult:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: boolean